tweak

+++ native/src/disktest.cpp
@@ -220,9 +220,9 @@ int main(int arg, char * param[])
       }
 
       long endAIO = getTime();
+ 
 
-
-      fprintf (stderr, "Total time to write file = %ld milliseconds, Bytes/millisecond = %ld, Writes/Syncs per millisecond = %ld \n", (endAIO - startAIO), (fileSize / (endAIO - startAIO)), (writes / (endAIO - startAIO)));
+      fprintf (stderr, "Total time to write file = %ld milliseconds, Bytes/millisecond = %ld, Writes/Syncs per millisecond = %f \n", (endAIO - startAIO), (fileSize / (endAIO - startAIO)), ((double)writes / ((double)endAIO - (double)startAIO)));
 
 
       free (events);
